chore: extract logic from api layer into own module
