+++ mobile/native/src/lib.rs
@@ -23,6 +23,7 @@ mod orderbook;
     unused_qualifications
 )]
 mod bridge_generated;
+mod channel_trade_constraints;
 mod cipher;
 mod destination;
 mod storage;
